Adding barcode trimming step

+++ BarcodeProcessingAndTrimming.sh
@@ -23,7 +23,7 @@ mkdir QC trim
 # QC raw reads
 #fastqc $R1 $R2 -t $CPU -o QC
 
-# Trim first NNN
+### Trim first NNN ###
 if test -f "trim/${R1trim}"; then
 	echo "File trim/${R1trim} already exist - SKIP TRIMMING"
 else
@@ -32,7 +32,7 @@ fi
 
 cd trim
 
-# BarcodeSplit
+### BarcodeSpliter ###
 mkdir BarSplitR1 BarSplitR2
 if test -f "BarSplitR1/${BARCODE1}_R1.fastq"; then
 	echo "Barcodes already splitted - SKIP BARCODESPLITTER"
@@ -41,9 +41,28 @@ else
  echo "zcat $R2trim | BarcodeSplitter --bcfile ${BARCODES} --bol --mismatches $MISMATCH --prefix BarSplitR2/ --suffix "_R2.fastq" --debug") | parallel
 fi
 
+
 conda activate fastq-pair # ! NEED fastq-pair conda enviroment with fastq_pair tool installed
 rm BarSplitR1/unmatched_R1.fastq BarSplitR2/unmatched_R2.fastq
+
 for barcode in `cat $BARCODES | cut -f1`
 	do
-	fastq_pair BarSplitR1/${barcode}_R1.fastq BarSplitR2/${barcode}_R2.fastq && rm BarSplitR1/${barcode}_R1.fastq BarSplitR2/${barcode}_R2.fastq
+	
+	### READ PAIRING ###
+	if test -f "BarSplitR1/${barcode}_R1.fastq.paired.fq"; then
+	        echo "Files already paired - SKIP PAIRING"
+	else
+		fastq_pair BarSplitR1/${barcode}_R1.fastq BarSplitR2/${barcode}_R2.fastq && rm BarSplitR1/${barcode}_R1.fastq BarSplitR2/${barcode}_R2.fastq
+	fi
+
+	### TRIM barcodes ###
+	if test -f "${barcode}_R1.trim.fq.gz"; then
+	
+		echo "Barcodes already trimmed - SKIP BARCODE TRIMMING"
+	
+	else	
+		fastp -w $CPU -f 22 -i BarSplitR1/${barcode}_R1.fastq.paired.fq -I BarSplitR1/${barcode}_R2.fastq.paired.fq -o ${barcode}_R1.trim.fq.gz -O ${barcode}_R2.trim.fq.gz
+
+	fi
+
 	done
